+++ Tests/JiraAPIClientTests/SignInTests.swift
@@ -22,7 +22,53 @@ final class SignInTests: JiraAPIClientTests {
         
         Mock(url: url, dataType: .json, statusCode: 200, data: [.post: data])
             .register()
-
+        
+        let resourcesURL = URL(string: "https://api.atlassian.com/oauth/token/accessible-resources")!
+        let resourcesData = """
+        [
+          {
+            "id": "1324a887-45db-1bf4-1e99-ef0ff456d421",
+            "name": "Site name",
+            "url": "https://your-domain.atlassian.net",
+            "scopes": [
+              "write:jira-work",
+              "read:jira-user",
+              "manage:jira-configuration"
+            ],
+            "avatarUrl": "https://site-admin-avatar-cdn.prod.public.atl-paas.net/avatars/240/flag.png"
+          }
+        ]
+        """
+            .data(using: .utf8)!
+        
+        Mock(url: resourcesURL, dataType: .json, statusCode: 200, data: [.get: resourcesData])
+            .register()
+        
+        let userProfileURL = URL(string: "https://api.atlassian.com/me")!
+        let userData = """
+        {
+          "account_type": "atlassian",
+          "account_id": "112233aa-bb11-cc22-33dd-445566abcabc",
+          "email": "mia@example.com",
+          "name": "Mia Krystof",
+          "picture": "https://avatar-management--avatars.us-west-2.prod.public.atl-paas.net/112233aa-bb11-cc22-33dd-445566abcabc/1234abcd-9876-54aa-33aa-1234dfsade9487ds",
+          "account_status": "active",
+          "nickname": "mkrystof",
+          "zoneinfo": "Australia/Sydney",
+          "locale": "en-US",
+          "extended_profile": {
+            "job_title": "Designer",
+            "organization": "mia@example.com",
+            "department": "Design team",
+            "location": "Sydney"
+          }
+        }
+        """
+            .data(using: .utf8)!
+        
+        Mock(url: userProfileURL, dataType: .json, statusCode: 200, data: [.get: userData])
+            .register()
+        
         cancellable = client.signIn()
             .sink(receiveCompletion: { completion in
                 switch completion {
@@ -35,7 +81,7 @@ final class SignInTests: JiraAPIClientTests {
                 signInFinished.fulfill()
             })
 
-        wait(for: [signInFinished], timeout: 1)
+        wait(for: [signInFinished], timeout: 5)
     }
 
     func testSignInFailsFromAuthenticationSession() {